requirements.txt
rdflib
pydantic
langchain
langchain-google-genai
google-generativeai
networkx
python-dotenv
pytest
streamlit
